requirements.txt
Flask==3.1.0
Flask-CORS==5.0.1
Flask-Mail==0.10.0
Werkzeug==3.1.3
pyjwt==2.10.1
bcrypt==4.3.0
python-dotenv==1.0.1
gunicorn==23.0.0
requests==2.32.3
Flask-Migrate==4.0.0
Flask-RESTful==1.0.0
Flask-RESTPlus==0.13.0
Flask-RESTX==1.0.0
Flask-Admin==1.6.0
Flask-SQLAlchemy==3.1.0
flask_jwt_extended